FUSETOOLS2-989 - use a specific name for each project test
without creating/recreating the 5 projects for each tests, even when
they are not used

execution time decreased from 4(number of unused project created per
test)x5(number of tests)xtime of project creation. So for me locally, it
reduced of 200 seconds
